move callbacks registering to ingredients controller
